+++ spine-ios/Example/Spine iOS Example/SpineExampleApp.swift
@@ -14,6 +14,35 @@ struct SpineExampleApp: App {
     init() {
         let version = Spine.version
         print("Spine \(version)")
+        
+        Task {
+            do {
+                // Load
+                
+                let atlas = try await Atlas.fromAsset("spineboy.atlas")
+                print(atlas)
+                
+                let skeletonDataFromJson = try SkeletonData.fromAsset(
+                    atlas: atlas,
+                    skeletonFile: "spineboy-pro.json"
+                )
+                print(skeletonDataFromJson)
+                
+                let skeletonDataFromBinary = try SkeletonData.fromAsset(
+                    atlas: atlas,
+                    skeletonFile: "spineboy-pro.skel"
+                )
+                print(skeletonDataFromBinary)
+                
+                // Dispose
+                
+                atlas.dispose()
+                skeletonDataFromJson.dispose()
+                skeletonDataFromBinary.dispose()
+            } catch {
+                print(error)
+            }
+        }
     }
     
     var body: some Scene {