+++ spine-ios/Sources/Spine/Spine.swift
@@ -1,7 +1,3 @@
-// The Swift Programming Language
-// https://docs.swift.org/swift-book
-
-import SpineSharedStructs
 import SpineWrapper
 
 public class Spine {