+++ spine-ios/CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add `spine-iOS` SPM package & example app (#1)
 - Basic Mesh Rendering (#2)
 - Spine C++ Swift Wrapper (#3)
+- Load `Atlas` & `SkeletonData` (#4)